+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2017-03-12
+* `[SEMVER_MINOR]` Second level transform specs can now be objects (will be wrapped by `createTransformer`)
+
 ## 2017-02-21
 * `[SEMVER_PATCH]` Added `json` config property to `createFileCache`